Upload a 2048 game :accept:

+++ simple-program/tools-in-python-shell/PY-to-MD-for-jekyll-POST.py
@@ -22,13 +22,16 @@ XXXX-XX-XX-!!!YOUR FILE NAME!!!
 import os, sys, re, shutil, time
 
 def trans(f):
-	filename = time.strftime("%Y-%m-%d-", time.localtime())+f[:-4]+'.MD'
-	if os.path.exists(filename):
-		os.remove(filename)
-	md = open(filename,'w', encoding='UTF-8')
-	all_the_text = open(f,'r', encoding='UTF-8').read()
-	briefing = re.compile("'''"+'(.*?)'+"'''",re.S).findall(all_the_text)
-	head = '\
+    if not os.path.exists(f):
+        return 0
+    print(f)
+    filename = time.strftime("%Y-%m-%d-", time.localtime())+f[:-3]+'.MD'
+    if os.path.exists(filename):
+        os.remove(filename)
+    md = open(filename,'w', encoding='UTF-8')
+    all_the_text = open(f,'r', encoding='UTF-8').read()
+    briefing = re.compile("'''"+'(.*?)'+"'''",re.S).findall(all_the_text)
+    head = '\
 ---\n\
 layout: postcn\n\
 title: "python:{}"\n\
@@ -41,28 +44,28 @@ tags: [python, file]\n\
 \n\
 * content\n\
 {}\n\n'.format(f[:-4], time.strftime("%Y-%m-%d", time.localtime()), time.strftime("%H:%M:%S", time.localtime()), '{:toc}')
-	md.write(head)
-	md.write(briefing[0])
-	md.write('<!-- more -->\n')
-	md.write('{% highlight ruby lineno %}\n')
-	md.write(all_the_text)
-	md.write('\n{% endhighlight %}')
-	md.close()
+    md.write(head)
+    md.write(briefing[0])
+    md.write('<!-- more -->\n')
+    md.write('{% highlight ruby lineno %}\n')
+    md.write(all_the_text)
+    md.write('\n{% endhighlight %}')
+    md.close()
+    return 1
 
 
 def checkpys():
-	i = 0
-	for root, dirs, files in os.walk(sys.path[0]):	#遍历当前文件夹
-		for f in files:
-			if '.py' in f:
-				trans(f)
-				i += 1
-	print('有{}个py文件被处理'.format(i))
+    i = 0
+    for root, dirs, files in os.walk(sys.path[0]):    #遍历当前文件夹
+        for f in files:
+            if '.py' in f:
+                i += trans(f)
+    input('有{}个py文件被处理'.format(i))
 
 #主执行程序开始
 print('当前操作目录：',sys.path[0])
 #myorder = input('输入任意字符开始，输入exit退出。注意本程序会将目录内所有py文件转化为MD文件！\n仅复制，源文件不会删除。您要：')
 #if myorder == "exit":
-#	sys.exit()
+#    sys.exit()
 print('------')
 checkpys()
